refactored project into modules
