+++ java/apirag-bedrock/src/main/java/com/datasqrl/ai/models/bedrock/BedrockFunctionCall.java
