+++ java/apirag-core/pom.xml
@@ -1,0 +1,23 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>com.datasqrl</groupId>
+    <artifactId>apirag</artifactId>
+    <version>0.1-SNAPSHOT</version>
+  </parent>
+  <artifactId>apirag-core</artifactId>
+  <name>apiRAG - Core Module</name>
+  <packaging>jar</packaging>
+
+  <properties>
+    <junit.jupiter.version>5.8.2</junit.jupiter.version>
+  </properties>
+
+
+  <dependencies>
+
+
+  </dependencies>
+</project>
