+++ java/apirag-core/src/main/java/com/datasqrl/ai/backend/ChatMessageInterface.java
