+++ java/apirag-core/src/main/java/com/datasqrl/ai/models/ChatClientProvider.java
@@ -1,0 +1,11 @@
+package com.datasqrl.ai.models;
+
+import java.util.List;
+import java.util.Map;
+
+public interface ChatClientProvider {
+
+  List<ResponseMessage> getChatHistory(Map<String, Object> context);
+
+  ResponseMessage chat(String message, Map<String, Object> context);
+}
