+++ java/apirag-core/src/main/java/com/datasqrl/ai/models/ChatMessageEncoder.java
