+++ java/apirag-core/src/main/java/com/datasqrl/ai/models/ResponseMessage.java
@@ -1,0 +1,18 @@
+package com.datasqrl.ai.models;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ResponseMessage {
+
+  private String role;
+  private String content;
+  private JsonNode chart;
+  private String uuid;
+  private String timestamp;
+
+
+}
