+++ java/apirag-openai/pom.xml
@@ -1,0 +1,54 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>com.datasqrl</groupId>
+    <artifactId>apirag</artifactId>
+    <version>0.1-SNAPSHOT</version>
+  </parent>
+  <artifactId>apirag-openai</artifactId>
+  <name>apiRAG - OpenAI Compatible Provider</name>
+  <packaging>jar</packaging>
+
+  <properties>
+    <openai-java.version>0.20.0</openai-java.version>
+    <jtokkit.version>0.6.1</jtokkit.version>
+    <aws.bedrock.version>2.25.45</aws.bedrock.version>
+  </properties>
+
+  <dependencies>
+    <dependency>
+      <groupId>com.datasqrl</groupId>
+      <artifactId>apirag-core</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+
+    <dependency>
+      <groupId>io.github.lambdua</groupId>
+      <artifactId>api</artifactId>
+      <version>${openai-java.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.github.lambdua</groupId>
+      <artifactId>service</artifactId>
+      <version>${openai-java.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.knuddels</groupId>
+      <artifactId>jtokkit</artifactId>
+      <version>${jtokkit.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.squareup.okhttp3</groupId>
+      <artifactId>logging-interceptor</artifactId>
+      <version>4.12.0</version>
+    </dependency>
+    <dependency>
+      <groupId>ai.djl.huggingface</groupId>
+      <artifactId>tokenizers</artifactId>
+      <version>${huggingface.version}</version>
+    </dependency>
+  </dependencies>
+
+</project>
