+++ java/apirag-openai/src/main/java/com/datasqrl/ai/models/groq/GroqChatModel.java
