+++ java/apirag-openai/src/main/java/com/datasqrl/ai/models/openai/OpenAiChatModel.java
