+++ java/apirag-server/src/main/java/com/datasqrl/ai/Examples.java
