+++ java/apirag-server/src/main/java/com/datasqrl/ai/ModelProvider.java
