+++ java/apirag-server/src/main/java/com/datasqrl/ai/PlotFunction.java
