+++ java/apirag-server/src/main/java/com/datasqrl/ai/spring/InputMessage.java
