+++ java/apirag-server/src/main/resources/plotfunction1d.json
