+++ java/apirag-server/src/main/resources/plotfunction2d.json
