+++ java/apirag-server/src/test/java/com/datasqrl/ai/CmdLineChatBot.java
